+++ src/main/java/com/example/darkshop/controllers/ImageController.java
@@ -19,7 +19,7 @@ public class ImageController {
     private final ImageRepository imageRepository;
 
     @GetMapping("/images/{id}")
-    private ResponseEntity<?> getImageById(@PathVariable Long id) {
+    private ResponseEntity<?> getImageById(@PathVariable Integer id) {
         Image image = imageRepository.findById(id).orElse(null);
         return ResponseEntity.ok()
                 .header("fileName", image.getOriginalFileName())
